Remove unintuitive return value for validExpirationWindow
The return value for validExpirationWindow was unintuitive.

An expiration window is valid when it is less than the certificate
validty period, because if the expiration window is greater than
the certificate validity period, then that certificate is *always*
scheduled for renewal.

validExpirationWindow returned true when the expiration_window was
invalid, and false when the expiration_window was valid. This is
backwards to normal expectations.

We never actually use this return value. It has no effect on the
subsequent flow of logic in the calling code, so it's less
confusing to just drop it.